RC Beam Design/requirements.txt
# CrewAI Framework - latest compatible version for Python 3.13
crewai>=1.7.2
crewai-tools>=0.19.7

# LangChain for RAG
langchain>=0.3.13
langchain-openai>=0.2.14
langchain-community>=0.3.15

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.10.5
numpy>=1.26.4

# RAG System Dependencies
chromadb>=0.5.23
pypdf>=5.1.0
